Refactor Status into a separate module
